pyproject.toml
[project]
version = "0.1.0"
name = "verbm"
description = "Language agnostic version bump tool"
requires-python = ">=3.9"
readme = "README.md"
license = { text = "Apache License (2.0)" }
authors = [{ name = "Maxim Voloshin", email = "mvolloshin@gmail.com" }]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
]

dependencies = ["PyYAML", "pydantic"]

[project.urls]
Homepage = "https://github.com/chocolacula/verbm"

[project.scripts]
verbm = "verbm.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
